requirements.txt
pdfplumber==0.11.9
flask==3.1.0
langchain==0.3.27
langchain-groq==0.3.1
requests==2.32.5
python-dotenv==1.0.0
markupsafe==2.1.3